% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d_clean.R
\name{d_clean}
\alias{d_clean}
\title{d_clean function}
\usage{
d_clean(df, var, file_type = "csv")
}
\arguments{
\item{df}{Dataframe that is not formatted, downloaded from Qualtrics.}

\item{var}{Name of the column serves as unique identifier in dataset.}

\item{file_type}{csv or xml}
}
\description{
This function takes dataframe and reshapes the data in panel format using unique identifying variable.
}
\examples{
csvdata <- read_csv("data/raw_data/DOSPERT_test.csv")
csvClean <- d_clean(csvdata, "uid", file_type = "csv")
xmldata <- xmlToDataFrame("data/raw_data/DOSPERT_test.xml", stringsAsFactors = F)
xmlClean <- d_clean(xmldata, "uid", file_type = "xml")
}
\keyword{dospert}

