% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d_score.R
\name{d_score}
\alias{d_score}
\title{d_score function}
\usage{
d_score(df, var, file_type = "csv")
}
\arguments{
\item{df}{Dataframe that is not formatted, downloaded from Qualtrics.}

\item{var}{Name of the column serves as unique identifier in dataset.}

\item{file_type}{csv or xml}
}
\description{
This function takes dataframe and appends the results of analyses at the left end of the original dataframe.
}
\examples{
csvdata <- read_csv("data/raw_data/DOSPERT_test.csv")
csvScore <- d_score(csvdata, "uid", file_type = "csv")
xmldata <- xmlToDataFrame("data/raw_data/DOSPERT_test.xml", stringsAsFactors = F)
xmlScore <- d_score(xmldata, "uid", file_type = "xml")
}
\keyword{dospert}

