% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d_sum.R
\name{d_sum}
\alias{d_sum}
\title{d_sum function}
\usage{
d_sum(df, var, domain, risk_type, file_type = "csv")
}
\arguments{
\item{df}{Dataframe that is not formatted, downloaded from Qualtrics.}

\item{var}{Name of the column serves as unique identifier in dataset.}

\item{domain}{Domain of interest}

\item{risk_type}{RT(risk taking), RB(risk benefit), RP(risk perception)}

\item{file_type}{csv or xml}
}
\description{
This function takes dataframe, variable that uniquely identifies the respondent, domain, risk type (RT, RB, RP) and file type (either csv or xml) to return the sum of the responses for the domain-risk type pair for each respondent.
}
\examples{
csvdata <- read_csv("data/raw_data/DOSPERT_test.csv")
finsum_csv <- d_sum(csvdata, "uid", "fin", "RT", file_type = "csv")
xmldata <- xmlToDataFrame("data/raw_data/DOSPERT_test.xml", stringsAsFactors = F)
finsum_xml <- d_sum(xmldata, "uid", "fin", "RT", file_type = "xml")
}
\keyword{dospert}

